Webgl / remove handling of element_index_uint extension
From now on we will assume this extension is always enabled.

An error message have been added in the unlikely scenario of a lack
of support.

+++ test/spec/ol/renderer/webgl/layer.test.js
@@ -89,7 +89,7 @@ describe('ol.renderer.webgl.Layer', function() {
 
     beforeEach(function() {
       vertexBuffer = new Float32Array(100);
-      indexBuffer = new Uint16Array(100);
+      indexBuffer = new Uint32Array(100);
       instructions = new Float32Array(100);
       elementIndex = 3;
